Reorganize components, Add openAI service and settings service
