+++ src/app/components/categories/categories.component.ts
