+++ src/app/components/search-trending/search-trending.component.ts
